Refactor + Multiple Pages
Not gonna lie, the multiple pages as it is is both ugly and useless, but hopefully it'll impress Leon. The real benefit to this commit is cutting down on all the redundency in ContinentButton.js
Also I just realized that I still have the form practice I was working on. This has no use, and should be gone. Whatever, I'll probably have a more final verson tomorrow morning.